requirements.txt
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
Flask-SocketIO==4.3.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pyOpenSSL==20.0.1
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
